+++ demo/sdl_opengl3/main.c
@@ -47,8 +47,8 @@
  * and the corresponding function. */
 /*#include "../style.c"*/
 /*#include "../calculator.c"*/
-/*#include "../overview.c" */
-/*#include "../node_editor.c" */
+/*#include "../overview.c"*/
+/*#include "../node_editor.c"*/
 
 /* ===============================================================
  *